Revert "attempting to fix coordinate system"
This reverts commit a56d286e8033b435c01a796edc88ac6ad9473555.

+++ gizmobot/giz_compass.h
@@ -10,7 +10,7 @@ class GizCompass{
 
 public:
   GizCompass();
-  double heading_r;
+  float heading_r;
   void update();
   void init();
 };
